requirements.txt
Flask==3.0.3
groq==0.11.0
python-docx==1.1.2
pdfplumber==0.11.4
python-dotenv==1.0.1